User Database Initialised.

+++ app/src/main/java/com/angel/shoppingapp/database/Basket.kt
@@ -9,5 +9,5 @@ data class Basket(
     val id: Int?,
     val image: String?,
     val title: String?,
-    val price: Double?
+    val price: Int?
 )
